requirements.txt
fastapi
uvicorn[standard]
openai
pdfplumber
python-multipart
python-dotenv
